Testing code commented plotting and speed test
The testing code has now been incoorporated into test/runtest.jl .
Due to some conflicts in the use of f and g, the test script
for predictor correcter now uses f_nondiag, g_nondiag.
Pkg.test on StochasticDiffEq passed.

A pull request will be submitted with this commit.
